+++ src/App.js
@@ -2,7 +2,6 @@ import React, {Component} from 'react';
 import {BrowserRouter, Route, Switch, Redirect} from "react-router-dom";
 import AdminLayout from "layouts/Admin.js";
 import LoginUI from "./components/Login/LoginUI";
-import ResponseUsers from "./radius-api/login-api/ResponseUsers";
 
 class App extends Component {
     render() {
@@ -16,10 +15,6 @@ class App extends Component {
                 {/*------------Public Route -------------------*/}
 
                 <Switch>
-
-
-
-                    <Route exact path='/data' component={ResponseUsers}/>
                     <Route path="/admin" render={(props) => <AdminLayout {...props} />}/>
                 </Switch>
             </BrowserRouter>